sizable

+++ src/main/java/com/amalitech/gpuconfigurator/dto/categoryconfig/CompatibleOptionDTO.java
@@ -7,5 +7,7 @@ import java.math.BigDecimal;
 
 @Builder
 public record CompatibleOptionDTO(String name, String type, CategoryConfig categoryConfig, BigDecimal price,
-                                  String media, String unit, Boolean isCompatible, Boolean isIncluded, Boolean isMeasured, Float baseAmount, Float maxAmount, Float priceIncrement) {
+                                  String media, String unit, Boolean isCompatible, Boolean isIncluded,
+                                  Boolean isMeasured, BigDecimal baseAmount, BigDecimal maxAmount,
+                                  BigDecimal priceIncrement) {
 }
