+++ src/main/java/com/amalitech/gpuconfigurator/dto/categoryconfig/CompatibleOptionResponseDto.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import java.math.BigDecimal;
 
 @Builder
-public record CompatibleOptionResponseDto(String id , String name, String type, BigDecimal price,
-                                  String media, String unit, Boolean isCompatible, Boolean isIncluded, Boolean isMeasured, Float priceIncrement, Float baseAmount, Float maxAmount) {
+public record CompatibleOptionResponseDto(String id, String name, String type, BigDecimal price,
+                                          String media, String unit, Boolean isCompatible, Boolean isIncluded,
+                                          Boolean isMeasured, BigDecimal priceIncrement, BigDecimal baseAmount,
+                                          BigDecimal maxAmount) {
 }
